+++ lib/ansible/modules/cloud/amazon/cloudtrail.py
@@ -245,7 +245,11 @@ trail:
 
 import traceback
 
-from botocore.exceptions import ClientError
+try:
+    from botocore.exceptions import ClientError
+except ImportError:
+    # Handled in main() by imported HAS_BOTO3
+    pass
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.ec2 import (boto3_conn, ec2_argument_spec, get_aws_connection_info,
@@ -257,7 +261,7 @@ def create_trail(module, client, ct_params):
     """
     Creates a CloudTrail
 
-    module : AnisbleModule object
+    module : AnsibleModule object
     client : boto3 client connection object
     ct_params : The parameters for the Trail to create
     """
@@ -274,7 +278,7 @@ def tag_trail(module, client, tags, trail_arn, curr_tags=None, dry_run=False):
     """
     Creates, updates, removes tags on a CloudTrail resource
 
-    module : AnisbleModule object
+    module : AnsibleModule object
     client : boto3 client connection object
     tags : Dict of tags converted from ansible_dict to boto3 list of dicts
     trail_arn : The ARN of the CloudTrail to operate on
@@ -400,7 +404,7 @@ def delete_trail(module, client, trail_arn):
     """
     Delete a CloudTrail
 
-    module : AnisbleModule object
+    module : AnsibleModule object
     client : boto3 client connection object
     trail_arn : Full CloudTrail ARN
     """
@@ -414,7 +418,7 @@ def update_trail(module, client, ct_params):
     """
     Delete a CloudTrail
 
-    module : AnisbleModule object
+    module : AnsibleModule object
     client : boto3 client connection object
     ct_params : The parameters for the Trail to update
     """